% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcloud2.R
\name{amcloud2}
\alias{amcloud2}
\title{Create wordcloud by amCharts.js}
\usage{
amcloud2(
  data,
  maxFontSize = 30,
  fontFamily = "Courier New",
  accuracy = 4,
  minWordLength = 2,
  maxCount = 200,
  steps = 15,
  rotation = 0,
  minColor = "#7f919e",
  maxColor = "#2b2b2b",
  excludeWords = NULL,
  bgColor = "#ffffff"
)
}
\arguments{
\item{data}{a dataframe, contains 2 columns named as 'tag' and 'weight'.}

\item{maxFontSize}{the size of the biggest words. It can be set in pixels, or in Percent; \strong{When using percent value, the lower the percent value the bigger the size} default to 30.}

\item{fontFamily}{character, default to 'Courier New'.}

\item{accuracy}{numeric, How accurate Word Cloud will be when searching for suitable place for a word. The higher the accuracy - the more closely packed word will be. \strong{Please note, higher accuracy comes at a cost of performance. Find your middle ground}. default to 4.}

\item{minWordLength}{numeric, default to 2.}

\item{maxCount}{numeric, default to 200.}

\item{steps}{numeric, default to 25.}

\item{rotation}{numeric, degree of rotation for the labels container.}

\item{minColor}{character, min color hex code (example: #7f919e).}

\item{maxColor}{character, max color hex code (example: #333333).}

\item{excludeWords}{character array, of words to exclude (example: c('word', 'to', 'remove')).}

\item{bgColor}{character, background color hex code, default to #ffffff.}
}
\description{
Function for Creating wordcloud from weighted words
}
\author{
Youness Bahi
}
